fix(project): `build_ext` in `tox` + remove `poetry`
Tox works by:

- Building a source distribution;
- Installing the source distribution to a virtualenv (testenv);
- Calls the desired commands (pytest and flake8 in our case).

Some `*.cpp` files don't exist when building a source distribution, they
should be generated from `*.pyx` files before, but `setuptools` deals
with this. The source distribution is created well.

**Problem 1**

The source distribution could not be built without Cython. A
build-dependency could be specified in `pyproject.toml`, but that is
parsed by `tox` only if `isolated_build = True`.

Using isolated build somehow doesn't work, we still have to run
`python setup.py build_ext --inplace` to create the shared libraries.
This could be hacked by overriding the default `pip install` command to
`pip install --no-binary piquasso {package}`, but that builds the shared
libraries into the isolated environment, not in the project. However
`pytest` runs in the project, where no shared libraries exist, resulting
in an `ImportError`.

_Solution_

We skip the sdist building, instead we install the package, and build
the Cython extension in-place. It is kind of hacky, but there is not
(yet) any nicer solution.

**Problem 2**

The package building is moved to `setup.py`, because `poetry` doesn't
offer much flexibility regarding the package building. But if `poetry`
isn't used for this, then `setup.py` should know about the installation
requirements.

Also, the `poetry`-specific part of `pyproject.toml` seems to be
understood by `pip`, but not by e.g. `conda`, which is bad.

Also also, the current solution of using `poetry` in `tox` is a mess,
IMHO.

_Solution_

The production dependencies are moved to `setup.py` under
`install_requires`, and the development dependencies in a separate
`requirements.txt`.

While doing this, `cython` and `flake8` got moved into
`requirements.txt` as a development dependency. Originally, they were
production dependencies, but they shouldn't be.

Not using `poetry` for building the package almost breaks all the other
features it could offer, therefore its job got separated to the
following packages:
```
pip  # Installing packages
twine  # Publishing packages
setuptools  # Building packages
virtualenv  # Virtualenvs, of course
```

We have lost the ability to lock the packages, but that even may be
irrelevant for multiple Python versions, per se.

Note, that our dependencies are disturbingly frozen, we'll have to deal
with that later.

The `README.md` and `.gitlab-ci.yml` got rewritten according to these
changes.

**Problem 3**

The `benchmarks/vacuum_homodyne` fails with `ImportError`. This
behaviour is the result of that we don't install `piquasso` with `pip` in our `.gitlab-ci.yml` currently.

_Solution_

The current folder `benchmarks` got appended to the `PYTHONPATH`.

References: #53