Multithreaded matrix product calculation
** Feature added **

The dot function is now parallelized following the strategy of divide-and-conquere.
The performance of the current method (for matrices typically used in the project) is in pair with the pure multithreaded numpy call.
(However, the performance highly fluctuate from run to run.)
Previously the calculation of matrix product was serial, it was parallelised only for larger matrices because of large parallelization overhead.
Now low level parallelization is applied, the serialization cutoff is lowered down to matrices relevant for the project.
To optimize for larger matrices further work needs to be done.

Adding unit test performance_tests/test_matrix.py for testing and banchmarking matrix multiplication with numpy.

setup.py:

Additionaly adding nonstandard TBB_INC_DIR and TBB_LIB_DIR library paths to the build.