update(piquasso): GaussianState
2 major things have been changed regarding `GaussianState`:

- The attributes `m`, `G`, `C` have been marked as private by naming
  them `_m`, `_G` and `_C`.

- The constructor does not accept these attributes as constructor
  arguments anymore. Other ways are made accessible to provide the mean
  and covariance matrix, instead.

These changes are made, since the underlying representation of the
Gaussian state should not be exposed to the user due to its tendency to
change. On the other hand, mainly the mean and covariance matrix is used
to characterize the Gaussian state anyway.

Regarding the `sampling` backend, the `ModelessOperation` (parent of
`Sampling`) has been removed, the regular `Operation` is used instead.
Moreover, the `Operation.params` attribute is refactored to be an
`OrderedDict` instead of a `tuple`.