No default searching path
It is possible that there is a TBB library on the default search path and the libraries would be linked with this, while
during the compilation different version of header files were included.
To avoid this, we omit the search of TBB on default paths, and make the TBB-devel python package as default.
Custom TBB libraries can be selected by TBB_LIB_DIR and TBB_INC_DIR environment variables. For details see the Readme.