feat: Threshold Boson Sampling realization
The input matrix for Gaussian Boson Sampling with threshold measurement
got replaced with matrices of real datatype.

The `TorontonianRecursive.cpp` expects the input matrix to be real.
Namely, if we omit the conversion to complex amplitudes, then the result
would be the same due to the nature of the Torontonian.

The `LAPACKE_dgetrf` and `LAPACKE_dgetri` are declared to be used in
`calc_Qinv`. The `calc_Qinv` is copied from the previous parent class of
`ThresholdBosonSampling` and changed to work with real matrices. The
`reduce` method got also created, which was essentially copied from
`GaussianState_Cov`.

The tests were updated accordingly.

Resolves #35