update(piquasso-module): Eliminating Circuit
The `piquasso-module` got updated. In it, all the `*Circuit` classes
were deleted, so it has to be followed up in `piquassoboost` as well.

Note, that the `result` and `shots` should be an attribute of a `State`.
It is not the best possible solution, but these attributes are needed to
be hardcoded in `GaussianState`, because it is subclassed from both the
C extension and the original Python class.