feat(Glynn permanent): Creating recursive permanent calculator
We are able to use as an advantage that the input matrix of
the permanent calculator algorithm often contains multiple rowes
or columns with same content. Using this we can create an algorithm
which can calculate the value of the permanent of such matrices
faster. Adding integer arrays to be able to perform compariosons on
integer values.

Tests were added with some cases.

This commit contains the skeleton of the class, the tests, the added
integer array skeleton.

Glynn permanent calculator was updated to be able to return valid result
if the given matrix is empty.