refactor(picstate): adding picstate with type int
This commit contains three parts:
1) Refactoring Glynn permanent calculator
2) Refactoring PicState datatype
3) Eliminating some warnings with new methods of matrix_base

Detailed:
1) Glynn permanent calculator is names GlynnPermanentCalculatorRepeated.h from now on,
since it is as recursive as the GlynnPermanenetCalculator. It uses the matrix multipled
data to be faster. The calculator class uses PicState to store the multiplicities of
the rows and columns instead of the array_int type. This means that class is not needed
anymore, so it id eliminated.

2) PicState datatype used int64_t to store the values. Due to performance reasons it
is changed to have the possibility to use datatype int to store the values. This meant
to create template class and general implementation of their methods. Convertation
function was created to be able to jump from PicState_int64 implementation to
PicState_int.

3) Due to constant and not constant objects of matrices and PicState's were used mixed,
the copy constructor and the operator[] was sometimes called on constant objects which
produced warning. Since there is possibility to avoid these kind of warnings constant
versions were implemented.