update: Piquasso 0.2.1 -> 0.5.0
**Seed**

The `piquasso.api.constants` module has been deleted in favour of
`pq.Config`. The usage of this module is deleted in PiquassoBoost.

Where the `constants` module is imported in some C++ Strategy class
wrapper, the required seed is passed in `__init__` instead.

There was a missing seed in `GaussianSimulationStrategy_wrapper`,
which got filled in.

Note, that the Python classes `GaussianSimulationStrategy` and
`GaussianSimulationStrategyFast` are used nowhere.

**GaussianState**

The `GaussianState` Python class with its wrapper and the
`CGaussianState` C++ class got deleted from PiquassoBoost, according to
the changes in `piquasso`'s API.

**Testing**

Since there is no `piquassoboost.patch`, the `_patch` fixture in
`conftest.py` is rewritten to "patch" the simulator classes only if
tests are run from the `piquasso-module` submodule.