feat(permanents): python interface for Glynn repeated permanent calculation
The permanent is calculated based on the input and output states which mean the
multiplicity of the specific rows and columns (e.g. (1..1) and (1..1) mean the
classical permanent calculation). This means the sum of input and output states
has to be equal.

Added python function for calculating the permanents with input and output
states of a matrix. If there are higher than 1 inputs or outputs the calculator
optimizes the calculation based on them.