refact(Permanent Calculators): templating
For being able to measure the precision of the permanent calculations, the glynn permanent calculator class were templated.
This means we can use the same code with different precision for the calculation.
Class aliases were added for the new calculators.
Python interface was updated.

New macro for the double precision calculation: GlynnDoubleCPU 6