feat(Permanent Calculators): Adding glynn calculator with simple algo
For measuring the accuracy of the permanent calculators, a new permanent calculator class was added with the simpliest
algorithm based simply on Glynn's formula (Balasubramanian-Bax-Franklin-Glynn (BBFG) formula)
The calculator class is templated to be able to calculate the permanent with different precisions.
Special template instances are defined in the piquassoboost/sampling/source/GlynnPermanentCalculatorSimple.cpp file.
Current available classes use double and long double precision