feat(BS lossy): Adding simulation strategy with uniform losses
A class for simulating uniform losses at BS was added with name GeneralizedCliffordsBUniformLossesSimulationStrategy.
This class uses the same algorithm as GeneralizedCliffordsBSimulationStrategy but with a specific concept:
it has a uniform probability whether it stays in the circuit or not.
This means the output particle number can differ from the input particle number (might decrease).
Additional wrapper class was added and python side implementation was extended with the uniform loss case.